+++ examples/dialog_save_state_example.py
@@ -2,8 +2,7 @@ import sys
 
 from qtpy import QtWidgets
 
-from eqt.ui import FormDialog
-from eqt.ui.UISliderWidget import UISliderWidget
+from eqt.ui import FormDialog, UISliderWidget
 
 
 class MainUI(QtWidgets.QMainWindow):
@@ -27,18 +26,25 @@ class MainUI(QtWidgets.QMainWindow):
         # ## Example on how to add elements to the
         dialog.addWidget(QtWidgets.QLabel('test label'), 'Label: ', 'label')
         dialog.addWidget(QtWidgets.QCheckBox('test checkbox'), 'CheckBox: ', 'checkBox')
+
         combobox = QtWidgets.QComboBox()
         combobox.addItems(['test1', 'test2'])
         dialog.addWidget(combobox, 'ComboBox: ', 'comboBox')
+
         dialog.addWidget(QtWidgets.QDoubleSpinBox(), 'DoubleSpinBox: ', 'doubleSpinBox')
         dialog.addWidget(QtWidgets.QSpinBox(), 'SpinBox: ', 'spinBox')
         dialog.addWidget(QtWidgets.QSlider(), 'Slider: ', 'slider')
-        dialog.addWidget(UISliderWidget(QtWidgets.QLabel()), 'UISliderWidget: ', 'uiSliderWidget')
-        dialog.addWidget(QtWidgets.QRadioButton('test 1'), 'RadioButton 1: ', 'radioButton')
-        dialog.addWidget(QtWidgets.QRadioButton('test 2'), 'RadioButton 2: ', 'radioButton_2')
+        dialog.addWidget(
+            UISliderWidget.UISliderWidget(minimum=-0.5, maximum=0.5, decimals=10,
+                                          number_of_steps=10, number_of_ticks=10),
+            'UISliderWidget: ', 'uiSliderWidget')
+
+        dialog.addWidget(QtWidgets.QRadioButton('test 1'), 'RadioButton 1: ', 'radioButton1')
+        dialog.addWidget(QtWidgets.QRadioButton('test 2'), 'RadioButton 2: ', 'radioButton2')
         dialog.addWidget(QtWidgets.QTextEdit('test'), 'TextEdit: ', 'textEdit')
         dialog.addWidget(QtWidgets.QPlainTextEdit('test'), 'PlainTextEdit: ', 'plainTextEdit')
         dialog.addWidget(QtWidgets.QLineEdit('test'), 'LineEdit: ', 'lineEdit')
+
         button = QtWidgets.QPushButton('test')
         button.setCheckable(True)
         dialog.addWidget(button, 'Button: ', 'button')